# Better font-finding heuristics, with a shortcut and a few caches.