--- a/src/precompile.jl
+++ b/src/precompile.jl
@@ -3,4 +3,9 @@ function _precompile_()
     @assert precompile(Tuple{typeof(findfont),String})   # time: 0.12886831
     @assert precompile(Tuple{typeof(try_load),String})   # time: 0.033520337
     @assert precompile(Tuple{typeof(renderface),FTFont,Char,Int64})   # time: 0.019107351
+    # Populate the font info cache
+    __init__()
+    for folder in fontpaths()
+        map(font_info, readdir(folder, join=true))
+    end
 end