@tecosaur 2024-02-01
Right. So it sounds like what we really want is a per-system font cache file?

Would `Scatch.jl` be the way to go then?